first commit before templatizing subrobot registration
