+++ src/robot/CMakeLists.txt
@@ -11,13 +11,9 @@ set(CMAKE_REQUIRED_FLAGS "")
 # Libraries
 #
 set(sources
-  ConcreteRobot.cpp
-  ConcreteManipulator.cpp
   GrabMetadata.cpp
-  Manipulator.cpp
   Robot.cpp
   util.cpp
-  GenericRobot.cpp
 )
 
 add_library("${PROJECT_NAME}_robot" SHARED ${sources})
