--- src/robot/GenericRobot.cpp
@@ -1,201 +1,0 @@
-#include "aikido/robot/GenericRobot.hpp"
-
-#include <dart/collision/fcl/FCLCollisionDetector.hpp>
-#include <srdfdom/model.h>
-#include <urdf/model.h>
-
-#include <aikido/constraint/TestableIntersection.hpp>
-#include <aikido/control/KinematicSimulationTrajectoryExecutor.hpp>
-#include <aikido/io/CatkinResourceRetriever.hpp>
-#include <aikido/planner/ConfigurationToConfiguration.hpp>
-
-namespace aikido {
-namespace robot {
-
-using constraint::dart::CollisionFreePtr;
-using statespace::dart::ConstMetaSkeletonStateSpacePtr;
-
-//==============================================================================
-GenericRobot::GenericRobot(
-    const dart::common::Uri& urdf,
-    const dart::common::Uri& srdf,
-    const std::string& name)
-  : mName(name)
-  , mCollisionDetector(dart::collision::FCLCollisionDetector::create())
-  , mSelfCollisionFilter(
-        std::make_shared<dart::collision::BodyNodeCollisionFilter>())
-{
-  // Read the URDF.
-  const dart::common::ResourceRetrieverPtr& retriever
-      = std::make_shared<io::CatkinResourceRetriever>();
-  dart::utils::DartLoader urdfLoader;
-  mMetaSkeleton = urdfLoader.parseSkeleton(urdf, retriever);
-  if (!mMetaSkeleton)
-  {
-    throw std::runtime_error("Unable to load the robot from URDF.");
-  }
-
-  // TODO: Avoid loading the URDF this second time. See PR63 [libherb].
-  urdf::Model urdfModel;
-  std::string urdfAsString = retriever->readAll(urdf);
-  urdfModel.initString(urdfAsString);
-
-  srdf::Model srdfModel;
-  std::string srdfAsString = retriever->readAll(srdf);
-  srdfModel.initString(urdfModel, srdfAsString);
-  auto disabledCollisions = srdfModel.getDisabledCollisionPairs();
-  for (auto disabledPair : disabledCollisions)
-  {
-    auto body0 = mMetaSkeleton->getBodyNode(disabledPair.link1_);
-    auto body1 = mMetaSkeleton->getBodyNode(disabledPair.link2_);
-    mSelfCollisionFilter->addBodyNodePairToBlackList(body0, body1);
-  }
-  mStateSpace = std::make_shared<statespace::dart::MetaSkeletonStateSpace>(
-      mMetaSkeleton.get());
-}
-
-//==============================================================================
-GenericRobot::GenericRobot(
-    const dart::dynamics::MetaSkeletonPtr& skeleton, const std::string& name)
-  : mName(name)
-{
-  mMetaSkeleton = skeleton;
-  mStateSpace = std::make_shared<statespace::dart::MetaSkeletonStateSpace>(
-      mMetaSkeleton.get());
-  // TODO(avk): Do we want to allow SRDF equivalence after construction?
-}
-
-//==============================================================================
-std::future<void> GenericRobot::executeTrajectory(
-    const trajectory::TrajectoryPtr& trajectory) const
-{
-  if (!mTrajectoryExecutor)
-  {
-    // TODO(avk): Should I try with the parent robot?
-    throw std::invalid_argument("Executor is null!");
-  }
-  return mTrajectoryExecutor->execute(trajectory);
-}
-
-//==============================================================================
-void GenericRobot::step(const std::chrono::system_clock::time_point& timepoint)
-{
-  // TODO(avk): Is the following statement valid? Where is it conveyed to the
-  // user and what are its implications?:
-  // Assumes that the parent robot is locked
-  mTrajectoryExecutor->step(timepoint);
-}
-
-// ==============================================================================
-CollisionFreePtr GenericRobot::getSelfCollisionConstraint(
-    statespace::dart::MetaSkeletonStateSpacePtr space,
-    dart::dynamics::MetaSkeletonPtr skeleton) const
-{
-  if (mRootRobot)
-  {
-    return mRootRobot->getSelfCollisionConstraint(space, skeleton);
-  }
-
-  // TODO(avk): Do I really need this post SRDF?
-  // TODO(avk): Can we move this to the constructor?
-  // TODO(avk): Why disable adjacent body check?
-  auto rootSkeleton = mMetaSkeleton->getBodyNode(0)->getSkeleton();
-  rootSkeleton->enableSelfCollisionCheck();
-  rootSkeleton->disableAdjacentBodyCheck();
-
-  // TODO: Switch to PRIMITIVE once this is fixed in DART.
-  // mCollisionDetector->setPrimitiveShapeType(FCLCollisionDetector::PRIMITIVE);
-  auto collisionOption
-      = dart::collision::CollisionOption(false, 1, mSelfCollisionFilter);
-  auto collisionFreeConstraint
-      = std::make_shared<constraint::dart::CollisionFree>(
-          space, skeleton, mCollisionDetector, collisionOption);
-  collisionFreeConstraint->addSelfCheck(
-      mCollisionDetector->createCollisionGroupAsSharedPtr(mMetaSkeleton.get()));
-  return collisionFreeConstraint;
-}
-
-//=============================================================================
-constraint::TestablePtr GenericRobot::getFullCollisionConstraint(
-    statespace::dart::MetaSkeletonStateSpacePtr space,
-    dart::dynamics::MetaSkeletonPtr skeleton,
-    const CollisionFreePtr& collisionFree) const
-{
-  if (mRootRobot)
-  {
-    return mRootRobot->getFullCollisionConstraint(
-        space, skeleton, collisionFree);
-  }
-
-  auto selfCollisionFree = getSelfCollisionConstraint(space, skeleton);
-  if (!collisionFree)
-  {
-    return selfCollisionFree;
-  }
-
-  // Make testable constraints for collision check.
-  std::vector<constraint::ConstTestablePtr> constraints{
-      selfCollisionFree, collisionFree};
-  return std::make_shared<constraint::TestableIntersection>(
-      mStateSpace, constraints);
-}
-
-//=============================================================================
-std::shared_ptr<GenericRobot> GenericRobot::registerChildRobot(
-    const dart::dynamics::MetaSkeletonPtr& metaSkeleton,
-    const std::string& name,
-    const RobotType type)
-{
-  if (mChildren.find(name) != mChildren.end())
-  {
-    throw std::invalid_argument("Child is already present");
-  }
-
-  // Dispatch via factory.
-  std::shared_ptr<GenericRobot> child;
-  if (type == RobotType::Manipulator)
-  {
-    child = std::make_shared<GenericManipulator>(metaSkeleton, name);
-  }
-  else if (type == RobotType::WheeledBase)
-  {
-    // TODO(avk): Defaulting to Manipulator again.
-    child = std::make_shared<GenericManipulator>(metaSkeleton, name);
-  }
-  // Update the root robot.
-  const auto& root = mRootRobot ? mRootRobot : this;
-  child->setRootRobot(root);
-  mChildren[name] = child;
-  return child;
-}
-
-//=============================================================================
-trajectory::TrajectoryPtr GenericRobot::planToConfiguration(
-    const planner::PlannerPtr& planner,
-    const statespace::StateSpace::State* goalState,
-    const constraint::TestablePtr& testableConstraint) const
-{
-  // Create the problem.
-  // TODO(avk): Get current state needs to be added.
-  auto problem = planner::ConfigurationToConfiguration(
-      mStateSpace, getCurrentState(), goalState, testableConstraint);
-
-  // Solve the problem with the provided planner.
-  return planner->plan(problem);
-}
-
-//=============================================================================
-std::shared_ptr<GenericRobot> GenericManipulator::registerChildRobot(
-    const dart::dynamics::MetaSkeletonPtr& metaSkeleton,
-    const std::string& name,
-    const RobotType type)
-{
-  auto child = GenericRobot::registerChildRobot(metaSkeleton, name, type);
-  if (type == RobotType::Gripper)
-  {
-    mGripper = std::dynamic_pointer_cast<GenericGripper>(child);
-  }
-}
-
-} // namespace robot
-} // namespace aikido
